Refactor ConferenceSchedule to for ease of traversal
This change modifies the ConferenceSchedule class so that the result of
presentations_by_time_slot is no longer a hash keyed on time slots with
values being an array of presentations. Instead, it returns an array of
TimeSlots, an object created to store the time and presentations that
are taking place at that time.